apps/dinov2-clustering/requirements.txt
streamlit
torch
torchvision
timm
pillow
numpy
matplotlib
scikit-learn
plotly
umap-learn
requests
